add results command
