+++ .k8s/commands.py
@@ -2,12 +2,10 @@
 import click
 from jinja2 import Template
 import utils
+import k8s
 from pathlib import Path
 import shutil
-from time import sleep
 import yaml
-from kubernetes.client.rest import ApiException
-import kubernetes
 
 # Variables
 templatesLocation = '.k8s/templates'
@@ -118,6 +116,16 @@ def test(framework, version, extra, namespace):
     results(framework, version, namespace)
 
 
+@click.command('results', short_help='Query results')
+@click.option('--framework', '-f', multiple=True, help="Framework to be tested.")
+@click.option('--version', '-v', multiple=True, help="Python version to be tested.")
+@click.option('--namespace', '-n', show_default=True, default="default", help="Run the in the specified namespace")
+def results(framework, version, namespace):
+    """Query the results for the given version and frameworks or filtered by them."""
+    click.echo(click.style(f"framework={framework} version={version}", fg='blue'))
+    k8s.results(framework, version, namespace)
+
+
 def deploy(framework, version, extra, namespace):
     """Given the python and framework then run the skaffold deployment"""
     # Enable the skaffold profiles matching the given framework and version, if any
@@ -129,46 +137,6 @@ def deploy(framework, version, extra, namespace):
         extraFlag = f'{extra}'
     command = f'skaffold deploy {extraFlag} --build-artifacts={generatedLocation}/tags.json -n {namespace} {profilesFlag}'
     utils.runCommand(command)
-
-
-def results(framework, version, namespace):
-    """Given the python and framework then gather the results when the jobs have finished"""
-    click.echo(click.style(f"TBC framework={framework} version={version}", fg='blue'))
-
-    ## Loop for each version/framework and activiley look for whether it has finished and if so the
-    #for ver in version:
-    #    for fram in framework:
-    # Configs can be set in Configuration class directly or using helper
-    # utility. If no argument provided, the config will be loaded from
-    # default location.
-    kubernetes.config.load_kube_config()
-    with kubernetes.client.ApiClient() as api_client:
-        api_instance = kubernetes.client.BatchV1Api(api_client)
-        try:
-            label_selector = f'repo=apm-agent-python,type=unit-test'
-            result = api_instance.list_namespaced_job(namespace, label_selector=label_selector)
-            #print(api_response)
-            print(len(result.items))
-            for r in result.items:
-                print(r.metadata.name)
-                get_job_status(api_instance, r.metadata.name, namespace)
-        except ApiException as e:
-            print("Exception when calling BatchV1Api->list_namespaced_job: %s\n" % e)
-
-
-def get_job_status(api_instance, job_name, namespace):
-    job_completed = False
-    while not job_completed:
-        api_response = api_instance.read_namespaced_job_status(
-            name=job_name,
-            namespace=namespace)
-        if api_response.status.succeeded is not None or \
-                api_response.status.failed is not None:
-            job_completed = True
-        sleep(1)
-        print('.', end=" ")
-        #print("Job status='%s'" % str(api_response.status))
-    print("Job '%s' finished" % str(job_name))
 
 
 def generateSkaffoldEntries(version, framework):
@@ -191,6 +159,7 @@ def generateSkaffoldEntries(version, framework):
 
     generateFrameworkProfiles(framework)
 
+
 def generateDefaultManifest(version):
     """Given the python then generate the default manifest"""
     output = defaultManifestTemplate.render(name=version, version=utils.getPythonVersion(version))
@@ -198,6 +167,7 @@ def generateDefaultManifest(version):
     profilesFile = f'{generatedLocation}/default.yaml'
     with open(profilesFile, 'w') as f:
         f.write(output)
+
 
 def generateVersionProfiles(version):
     """Given the python then update the generated skaffold profiles for that version"""
